ai_service/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
mediapipe
numpy
Pillow
python-multipart
pydantic
